+++ messagingpush/src/main/java/io/customer/messagingpush/CustomerIOPushReceiver.kt
@@ -10,9 +10,9 @@ import io.customer.messagingpush.CustomerIOPushNotificationHandler.Companion.DEE
 import io.customer.messagingpush.CustomerIOPushNotificationHandler.Companion.DELIVERY_ID
 import io.customer.messagingpush.CustomerIOPushNotificationHandler.Companion.DELIVERY_TOKEN
 import io.customer.messagingpush.CustomerIOPushNotificationHandler.Companion.NOTIFICATION_REQUEST_CODE
+import io.customer.messagingpush.data.request.MetricEvent
 import io.customer.sdk.CustomerIO
 import io.customer.sdk.CustomerIOConfig
-import io.customer.sdk.data.request.MetricEvent
 import io.customer.sdk.di.CustomerIOComponent
 
 class CustomerIOPushReceiver : BroadcastReceiver() {
@@ -41,7 +41,7 @@ class CustomerIOPushReceiver : BroadcastReceiver() {
         val deliveryToken = bundle?.getString(DELIVERY_TOKEN)
 
         if (deliveryId != null && deliveryToken != null) {
-            CustomerIO.instance().trackMetric(deliveryId, MetricEvent.opened, deliveryToken)
+            MessagingPush.instance().trackMetric(deliveryId, MetricEvent.opened, deliveryToken)
         }
 
         val deepLink = bundle?.getString(DEEP_LINK_KEY)
