+++ messagingpush/src/main/java/io/customer/messagingpush/data/request/Device.kt
@@ -1,17 +1,17 @@
-package io.customer.sdk.data.request
+package io.customer.messagingpush.data.request
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import java.util.*
 
 @JsonClass(generateAdapter = true)
-data class Device(
+internal data class Device(
     @field:Json(name = "id") val token: String,
     val platform: String = "android",
     val lastUsed: Date
 )
 
 @JsonClass(generateAdapter = true)
-data class DeviceRequest(
+internal data class DeviceRequest(
     val device: Device
 )
