+++ messagingpush/src/main/java/io/customer/messagingpush/data/request/Metric.kt
@@ -1,4 +1,4 @@
-package io.customer.sdk.data.request
+package io.customer.messagingpush.data.request
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
@@ -10,7 +10,7 @@ enum class MetricEvent {
 }
 
 @JsonClass(generateAdapter = true)
-data class Metric(
+internal data class Metric(
     @field:Json(name = "delivery_id") val deliveryID: String,
     @field:Json(name = "device_id") val deviceToken: String,
     val event: MetricEvent,
