+++ messagingpush/src/main/java/io/customer/messagingpush/queue/taskdata/DeletePushNotificationQueueTaskData.kt
@@ -3,7 +3,7 @@ package io.customer.messagingpush.queue.taskdata
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class DeletePushNotificationQueueTaskData(
+internal data class DeletePushNotificationQueueTaskData(
     val profileIdentified: String,
     val deviceToken: String
 )
