+++ messagingpush/src/main/java/io/customer/messagingpush/queue/taskdata/RegisterPushNotificationQueueTaskData.kt
@@ -4,7 +4,7 @@ import java.util.*
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class RegisterPushNotificationQueueTaskData(
+internal data class RegisterPushNotificationQueueTaskData(
     val profileIdentified: String,
     val deviceToken: String,
     val lastUsed: Date
