+++ messagingpush/src/main/java/io/customer/messagingpush/queue/type/QueueTaskType.kt
@@ -3,5 +3,6 @@ package io.customer.messagingpush.queue.type
 // All the types of tasks the MessagingPush module runs in the background queue
 internal enum class QueueTaskType {
     RegisterDeviceToken,
-    DeletePushToken
+    DeletePushToken,
+    TrackPushMetric
 }
