+++ sdk/src/main/java/io/customer/sdk/queue/type/QueueTaskType.kt
@@ -3,6 +3,5 @@ package io.customer.sdk.queue.type
 // All the types of tasks the Tracking module runs in the background queue
 internal enum class QueueTaskType {
     IdentifyProfile,
-    TrackEvent,
-    TrackPushMetric
+    TrackEvent
 }
